wip, before full rename
more fixes, revert some changes

more refactorings

minor

arch to tt

fix after rebase

revert Arch to ARCH
