+++ device/api/umd/device/architecture_implementation.h
@@ -12,9 +12,9 @@
 #include <vector>
 
 #include "umd/device/tlb.h"
-#include "umd/device/tt_arch_types.h"
 #include "umd/device/tt_xy_pair.h"
-#include "umd/device/xy_pair.h"
+#include "umd/device/types/arch.h"
+#include "umd/device/types/xy_pair.h"
 
 struct tt_driver_host_address_params;
 struct tt_driver_eth_interface_params;
