+++ device/api/umd/device/coordinate_manager.h
@@ -10,9 +10,9 @@
 #include <set>
 #include <vector>
 
-#include "umd/device/tt_arch_types.h"
 #include "umd/device/tt_core_coordinates.h"
 #include "umd/device/tt_xy_pair.h"
+#include "umd/device/types/arch.h"
 
 class CoordinateManager {
 public:
