+++ device/api/umd/device/tt_xy_pair.h
@@ -8,7 +8,8 @@
 
 #include <regex>
 
-#include "umd/device/xy_pair.h"
+#include "umd/device/types/xy_pair.h"
 
+// TODO: These usings and this whole file are to be removed.
 using tt_xy_pair = tt::umd::xy_pair;
 using tt_cxy_pair = tt::umd::cxy_pair;
