+++ tests/api/test_mockup_device.cpp
@@ -12,7 +12,7 @@
 
 #include "device/mockup/tt_mockup_device.hpp"
 #include "tests/test_utils/generate_cluster_desc.hpp"
-#include "umd/device/tt_arch_types.h"
+#include "umd/device/types/arch.h"
 
 namespace test::mockup_device {
 
@@ -22,23 +22,6 @@ std::string get_env_arch_name() {
         return arch_name_ptr;
     }
     throw std::runtime_error("Environment variable ARCH_NAME is not set.");
-}
-
-tt::ARCH get_arch_from_string(const std::string &arch_str) {
-    if (arch_str == "grayskull" || arch_str == "GRAYSKULL") {
-        return tt::ARCH::GRAYSKULL;
-    }
-    if (arch_str == "wormhole_b0" || arch_str == "WORMHOLE_B0") {
-        return tt::ARCH::WORMHOLE_B0;
-    }
-    if (arch_str == "blackhole" || arch_str == "BLACKHOLE") {
-        return tt::ARCH::BLACKHOLE;
-    }
-    if (arch_str == "Invalid" || arch_str == "INVALID") {
-        return tt::ARCH::Invalid;
-    }
-
-    throw std::runtime_error(arch_str + " is not recognized as tt::ARCH.");
 }
 
 std::string get_soc_descriptor_file(tt::ARCH arch) {
@@ -59,7 +42,7 @@ std::string get_soc_descriptor_file(tt::ARCH arch) {
 }
 
 TEST(ApiMockupTest, CreateDevice) {
-    const auto arch = get_arch_from_string(get_env_arch_name());
+    const auto arch = tt::arch_from_str(get_env_arch_name());
     std::cout << "Creating mockup device" << std::endl;
     auto device_driver = std::make_unique<tt_MockupDevice>(get_soc_descriptor_file(arch));
 }
